Move the args canon gate beside the type it specifies
`mod args_canon` tests five error families against the `ERROR.md` table;
one of them lived in the host file, and the type under contract —
`Diagnostic::args`, `with_args`, `diag_args!` — lived in `error.rs`. The
address said validation owned a cross-cutting rule, so a sixth family
growing an `args()` had no reason to land there.

Tests are unchanged: the four absolute-path imports stay, `ValidationError`
becomes a `crate::quill::` import like its siblings, and `include_str!`
climbs three levels instead of four. `ERROR.md` § Coverage names the new
path.
