+++ crates/core/src/error.rs
@@ -668,3 +668,238 @@ mod tests {
     }
 
 }
+
+/// The canon table in `prose/canon/ERROR.md` § "Diagnostic args" is the contract
+/// a consumer writes its string table against, so it is tested like one rather
+/// than maintained by hand beside the code.
+#[cfg(test)]
+mod args_canon {
+    use std::collections::BTreeMap;
+
+    use super::ParseError;
+    use crate::document::EditError;
+    use crate::quill::{CoercionError, ValidationError};
+
+    /// `code` → its arg keys, sorted. Every variant of every enum on the
+    /// structured surface appears once.
+    fn minted() -> BTreeMap<String, Vec<String>> {
+        let mut out: BTreeMap<String, Vec<String>> = BTreeMap::new();
+        let mut add = |code: &str, args: BTreeMap<String, serde_json::Value>| {
+            let keys: Vec<String> = args.keys().cloned().collect();
+            assert!(
+                out.insert(code.to_string(), keys).is_none(),
+                "two samples for `{code}`: one code carries one payload"
+            );
+        };
+
+        for e in [
+            ValidationError::TypeMismatch {
+                path: "main.n".into(),
+                expected: "string".into(),
+                actual: "integer".into(),
+                source_token: "42".into(),
+                default: Some("\"x\"".into()),
+            },
+            ValidationError::EnumViolation {
+                path: "main.tone".into(),
+                value: "loud".into(),
+                allowed: vec!["quiet".into()],
+            },
+            ValidationError::FormatViolation {
+                path: "main.when".into(),
+                format: "date".into(),
+            },
+            ValidationError::UnknownCard {
+                path: "cards[0]".into(),
+                card: "ghost".into(),
+            },
+            ValidationError::BodyDisabled {
+                path: "cards.sig[0].body".into(),
+                card: "sig".into(),
+            },
+            ValidationError::NotInline {
+                path: "main.title".into(),
+            },
+            ValidationError::NotPlain {
+                path: "main.title".into(),
+            },
+        ] {
+            add(e.code(), e.args());
+        }
+
+        for e in [
+            EditError::InvalidFieldName("9bad".into()),
+            EditError::UnknownField("nope".into()),
+            EditError::InvalidKindName("Bad".into()),
+            EditError::ReservedKind,
+            EditError::IndexOutOfRange { index: 3, len: 1 },
+            EditError::ValueTooDeep { max: 8 },
+            EditError::Import(quillmark_content::import::ImportError::NestingTooDeep {
+                depth: 9,
+                max: 8,
+            }),
+            EditError::FieldRichtextDecode {
+                field: "body".into(),
+                message: "x".into(),
+            },
+            EditError::FieldNotContent {
+                field: "qty".into(),
+                declared: "integer".into(),
+            },
+            EditError::FieldRichtextNotInline("body".into()),
+            EditError::FieldConform {
+                field: "n".into(),
+                target: "integer".into(),
+                message: "x".into(),
+            },
+            EditError::ContentApply(quillmark_content::ApplyError::LineOutOfRange {
+                line: 3,
+                lines: 1,
+            }),
+        ] {
+            add(e.code(), e.args());
+        }
+
+        // The `conform::*` family: the strict write's refusals, re-namespaced by
+        // `conform_diagnostic`. Minted through that function rather than
+        // re-derived, so the table cannot drift from the code that stamps it.
+        for e in [
+            EditError::InvalidFieldName("9bad".into()),
+            EditError::ValueTooDeep { max: 8 },
+            EditError::FieldRichtextNotInline("body".into()),
+            EditError::FieldRichtextDecode {
+                field: "body".into(),
+                message: "x".into(),
+            },
+            EditError::FieldConform {
+                field: "n".into(),
+                target: "integer".into(),
+                message: "x".into(),
+            },
+        ] {
+            let diag = crate::quill::conform::conform_diagnostic(&e, &crate::DocPath::main());
+            add(
+                diag.code.as_deref().expect("conform diagnostics carry a code"),
+                diag.args,
+            );
+        }
+
+        for e in [
+            ParseError::InputTooLarge { size: 2, max: 1 },
+            ParseError::InvalidStructure("x".into()),
+            ParseError::EmptyInput("x".into()),
+            ParseError::MissingQuill("x".into()),
+            ParseError::BodyImport("x".into()),
+            ParseError::InvalidQuillReference {
+                value: "a@b".into(),
+                reason: "x".into(),
+            },
+            ParseError::YamlErrorWithLocation {
+                message: "x".into(),
+                line: 3,
+                block_index: 1,
+                hint: None,
+            },
+        ] {
+            let diag = e.to_diagnostic();
+            add(diag.code.as_deref().expect("parse errors carry a code"), diag.args);
+        }
+
+        // Two codes are minted beside their error rather than from a variant:
+        // `compose::coercion_error` wraps the whole `CoercionError`, and
+        // `compose::fill_warning` has no error type at all.
+        add(
+            "validation::coercion_failed",
+            CoercionError::Uncoercible {
+                path: "card_kinds.sig.n".into(),
+                value: "\"x\"".into(),
+                target: "integer".into(),
+                reason: "string is not a valid integer".into(),
+            }
+            .args(),
+        );
+        add("validation::must_fill", BTreeMap::new());
+
+        out
+    }
+
+    /// The `| code | args | outcome |` rows of the canon table, keyed the same
+    /// way. `—` is no keys; a trailing `?` marks a conditional key, which the
+    /// sample above supplies.
+    fn declared() -> BTreeMap<String, Vec<String>> {
+        let canon = include_str!("../../../prose/canon/ERROR.md");
+        let mut rows = canon
+            .lines()
+            .skip_while(|l| !l.starts_with("| Code | Args | Outcome |"))
+            .skip(2)
+            .take_while(|l| l.starts_with('|'));
+
+        let mut out = BTreeMap::new();
+        for row in &mut rows {
+            let cells: Vec<&str> = row.trim_matches('|').split('|').map(str::trim).collect();
+            assert_eq!(cells.len(), 3, "malformed canon row: {row}");
+            let code = cells[0].trim_matches('`').to_string();
+            let keys = if cells[1] == "—" {
+                Vec::new()
+            } else {
+                let mut keys: Vec<String> = cells[1]
+                    .split(',')
+                    .map(|k| k.trim().trim_end_matches('?').trim_matches('`').to_string())
+                    .collect();
+                keys.sort();
+                keys
+            };
+            assert!(out.insert(code, keys).is_none(), "duplicate canon row: {row}");
+        }
+        assert!(!out.is_empty(), "canon args table not found in ERROR.md");
+        out
+    }
+
+    /// The other direction: a code off the table carries no args, so a
+    /// consumer's template falls back rather than half-filling. `quill::*` is
+    /// the largest such family and the one with `format!`-built codes.
+    #[test]
+    fn out_of_scope_codes_carry_no_args() {
+        let diags = crate::quill::QuillConfig::from_yaml_with_warnings(
+            r#"
+Quill:
+  name: t
+  version: "1.0"
+  backend: typst
+  description: A slot whose literal contradicts its declared type
+
+main:
+  fields:
+    title:
+      type: string
+      default: 42
+"#,
+        )
+        .expect_err("a default that contradicts its type fails config validation");
+
+        assert!(
+            diags.iter().any(|d| d
+                .code
+                .as_deref()
+                .is_some_and(|c| c.starts_with("quill::"))),
+            "expected a quill:: diagnostic, got {:?}",
+            diags.iter().map(|d| &d.code).collect::<Vec<_>>()
+        );
+        for d in &diags {
+            assert!(
+                d.args.is_empty(),
+                "`{:?}` is off the canon table and must carry no args",
+                d.code
+            );
+        }
+    }
+
+    #[test]
+    fn diagnostic_args_match_canon() {
+        assert_eq!(
+            declared(),
+            minted(),
+            "`ERROR.md` § \"Diagnostic args\" and the minted args disagree"
+        );
+    }
+}
