+++ prose/canon/ERROR.md
@@ -246,7 +246,7 @@ Values keep their JSON shape (`allowed` arrives as a list, `len` as a number) be
 
 ### Coverage
 
-The table is the contract, so it is tested like one: `diagnostic_args_match_canon` in `crates/core/src/quill/validation.rs` fails if code and canon disagree, and its twin asserts the codes off the table carry nothing. Each `args()` binds every field rather than eliding with `..`, so a new field on a variant does not compile until it decides whether it is message-relevant.
+The table is the contract, so it is tested like one: `diagnostic_args_match_canon` in `crates/core/src/error.rs` fails if code and canon disagree, and its twin asserts the codes off the table carry nothing. Each `args()` binds every field rather than eliding with `..`, so a new field on a variant does not compile until it decides whether it is message-relevant.
 
 Three outcomes, and the wire tells them apart only with this table in hand, since "no keys" reads identically for the second and third:
